Added a simple rule to detect eml attachments
Added a simple rule that grabs any attachment with the extension `eml` and sends it to binexplode. Perhaps there should be an `or` condition instead of an `and` condition for matches in the `binexplode` on line 11. Another `or` statement to find instances using the yara flavor may be beneficial as well.

I couldn't pick a tag that seemed relevant, nor is there an "info" severity, so I omitted both of those.