Add WIP workflow for updating test rules
This is just the first step -- triggering the job, checkout
out the branch, and validating that the basic validation
workflow completed successfully.

I've validated these steps, but need to figure out some behavior
related to forks. To do that testing I'm bringing this int
sublime-rules so I can fork and test.