Update link_pikabot_smb_feb24.yml
Hey @padey  thank you!! 

Updating to use the url.scheme to determine if the link is a file:// 
And because you're already calling any(body.links once, I reduced the second any(body.links call to be an and condition to check the domain.domain in $abuse_ch_urlhaus_domains_trusted_reporters.

Fundamentally i think the logic is good here, the issue I'm hitting in testing is that because the link is to a file, we only produce a url & path in the MDM. Meaning this rule would never match on the domain.domain, nor would link analysis ever execute 

https://analyzer.sublimesecurity.com?id=a4b58d9b-2819-419a-a5d9-f3a5ebaeefe6

If you look at the link above, you'll see in the MDM by expanding body>links, that the scheme:file only produces a url and a path. But the URL doesn't get parsed further than that(likely due to the file schema). Which means we have no .domain.domain to check against, or a proper link to hand to linkanalysis.